refactor: split oversized files to satisfy max-lines ESLint rule
Extract FamilyStatistics interface to lib/statistics/types.ts and
split achievement-progress-service into types/evaluators modules,
bringing both files under the 300-line limit and removing their
eslint waivers. Split three oversized test files (service-routing,
service-upsert, ClassChangeForm) by extracting logical groups into
separate files. Fix pre-existing missing supabase import in
useRewardStoreActions.ts.
