+++ components/profile/ClassChangeForm.test.tsx
@@ -271,56 +271,4 @@ describe("ClassChangeForm", () => {
       expect(screen.getByText("Server error")).toBeInTheDocument();
     });
   });
-
-  it("falls back to refreshSession when getSession returns no token", async () => {
-    const { supabase: mockSupabase } = jest.requireMock("@/lib/supabase");
-    (mockSupabase.auth.getSession as jest.Mock).mockResolvedValueOnce({
-      data: { session: null },
-    });
-    global.fetch = jest.fn().mockResolvedValue({
-      ok: true,
-      json: () => Promise.resolve({}),
-    });
-
-    const user = userEvent.setup();
-    render(
-      <ClassChangeForm character={mockCharacter} onSuccess={mockOnSuccess} />,
-    );
-
-    const knightButton = await screen.findByRole("button", { name: /Knight/i });
-    await user.click(knightButton);
-    await user.click(
-      screen.getByRole("button", { name: /Confirm Class Change/i }),
-    );
-    await user.click(screen.getByRole("button", { name: /^Change Class$/i }));
-
-    await waitFor(() => {
-      expect(mockSupabase.auth.refreshSession).toHaveBeenCalled();
-      expect(global.fetch).toHaveBeenCalledWith(
-        expect.stringContaining("/change-class"),
-        expect.objectContaining({
-          headers: expect.objectContaining({
-            Authorization: "Bearer refreshed-token",
-          }),
-        }),
-      );
-    });
-  });
-
-  it("formats cooldown timer correctly", async () => {
-    (ProfileService.canChangeClass as jest.Mock).mockResolvedValue(false);
-    (
-      ProfileService.getClassChangeCooldownRemaining as jest.Mock
-    ).mockResolvedValue(
-      86400000 + 3600000 + 300000, // 1d 1h 5m
-    );
-
-    render(
-      <ClassChangeForm character={mockCharacter} onSuccess={mockOnSuccess} />,
-    );
-
-    await waitFor(() => {
-      expect(screen.getByText(/1d 1h 5m/)).toBeInTheDocument();
-    });
-  });
 });
